Solution to Assignment 4

+++ kern/include/fsystemcalls.h
@@ -13,13 +13,13 @@ int open(const char *, int, int32_t *);
 
 int write(int, userptr_t, size_t, int32_t *);
 
-int read(int, userptr_t, size_t , int32_t *);
+int read(int, userptr_t, size_t , ssize_t *);
 
 int lseek(int, off_t, int , int32_t*, int32_t *);
 
 int close(int);
 
-int dup2(int, int, int*);
+int dup2(int, int, int32_t*);
 
 int chdir(const char*);
